improving annotations and new modify Blender module
