+++ core/__init__.py
@@ -13,7 +13,8 @@ core_modules = [
     "monad_properties", "sv_custom_exceptions",
     "node_id_dict", "links", "sockets",
     "handlers", "update_system", "upgrade_nodes",
-    "monad", "node_defaults", "events"
+    "monad", "node_defaults", "events",
+    "modify_blender"
 ]
 
 def sv_register_modules(modules):
